still refactoring submitFunc
