+++ src/css/events.css
@@ -9,3 +9,6 @@
 .active-block{
     display: block;
 }
+
+
+
