+++ src/helpers/selectComments.ts
@@ -1,5 +1,5 @@
 import Store from "../models/Store";
-import commentInterface from "../interfaces";
+import commentInterface from "../interfaces/IComment";
 import saveProperComments from "./saveProperComments";
 
 export default function selectComments(store: Store, json: commentInterface[]): void {
