+++ store.html
@@ -4,7 +4,6 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>LIBRA | Store</title>
         <link href="src/css/styleSheet.css" rel="stylesheet" type="text/css"/>
-        <link href="form.css" rel="stylesheet" type="text/css"/>
         <link href="src/css/events.css" rel="stylesheet" type="text/css"/>
     
     </head>
